serverless/pyproject.toml
[project]
name = "serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
